Servers removed from server selection list on the profile admin page: *Fairy *Kujata *Remora *Midgardsormr *Hades *Seraph *Garuda *Pandemonium
Added automatic update to update this field in your database so you don't have to worry about doing it. For this to take effect, you must disable and reenable the plugin. I'm sorry, there is no way around this. You should be doing this on all plugin updates anyway.

Found a typo in admin/preview.php and corrected.
Found a typo in ffxi.php and corrected.

In admin/profile.php
Changed $world to $server
Altered the checked and selected functions incoming variable names.